Podcast List Block
Creates a dynamic Podcast List block.
Refactors the Player Code into it's own class.
