+++ seriously-simple-podcasting.php
@@ -23,11 +23,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+use SeriouslySimplePodcasting\Blocks\Castos_Blocks;
 use SeriouslySimplePodcasting\Controllers\Admin_Controller;
 use SeriouslySimplePodcasting\Controllers\Frontend_Controller;
 use SeriouslySimplePodcasting\Controllers\Settings_Controller;
 use SeriouslySimplePodcasting\Controllers\Options_Controller;
-use SeriouslySimplePodcasting\Blocks\Player_Block;
 use SeriouslySimplePodcasting\Rest\Rest_Api_Controller;
 
 define( 'SSP_VERSION', '2.1.0' );
@@ -70,7 +70,7 @@ if ( is_admin() ) {
 /**
  * Blocks
  */
-$ssp_player_block = new Player_Block(__FILE__, SSP_VERSION);
+new Castos_Blocks(__FILE__, SSP_VERSION);
 /**
  * Only load WP REST API Endpoints if the WordPress version is newer than 4.7
  */
