+++ src/components/PodcastList.js
@@ -1,0 +1,12 @@
+/**
+ * WordPress dependencies
+ */
+import {Component} from '@wordpress/element';
+
+class PodcastList extends Component {
+	render(){
+		return <div>Podcast List</div>
+	}
+}
+
+export default PodcastList;
